more pointers arrays

+++ 0x06-pointers_arrays_strings/main.h
@@ -5,7 +5,7 @@
 #include <string.h>
 
 char *_strcat(char *dest, char *src);
-char *_strncat(char *dest, char *src, int n);
+char *_strncat(char *dest, char *src, unsigned long n);
 char *_strncpy(char *dest, char *src, int n);
 int _strcmp(char *s1, char *s2);
 void reverse_array(int *a,int n);
